Refactor: EntityDso < EntityDto < IEntityRepo_H.

+++ plover/modeling/plover-orm/src/main/java/com/bee32/plover/orm/entity/MapEntityRepository.java
@@ -92,7 +92,7 @@ public class MapEntityRepository<E extends IEntity<K>, K extends Serializable>
     }
 
     @Override
-    public long count() {
+    public int count() {
         return getMap().size();
     }
